refactor(scripting): inline getChainDeploymentInfo into example script
Remove the standalone SDK function and schema. The example script now
composes existing SDK primitives directly, matching the pattern used
by deployNewChain and createTokenBridge examples.
