+++ src/index.ts
@@ -131,11 +131,6 @@ import {
 import * as utils from './utils';
 
 import { getBridgeUiConfig, GetBridgeUiConfigFunctionParams } from './getBridgeUiConfig';
-import {
-  getChainDeploymentInfo,
-  GetChainDeploymentInfoParams,
-  ChainDeploymentInfo,
-} from './getChainDeploymentInfo';
 import { getDefaultConfirmPeriodBlocks } from './getDefaultConfirmPeriodBlocks';
 import { getDefaultChallengeGracePeriodBlocks } from './getDefaultChallengeGracePeriodBlocks';
 import { getDefaultMinimumAssertionPeriod } from './getDefaultMinimumAssertionPeriod';
@@ -315,10 +310,6 @@ export {
   getBridgeUiConfig,
   BridgeUiConfig,
   GetBridgeUiConfigFunctionParams,
-  //
-  getChainDeploymentInfo,
-  GetChainDeploymentInfoParams,
-  ChainDeploymentInfo,
   //
   getValidators,
   GetValidatorsParams,
