+++ src/scripting/schemaCoverage.unit.test.ts
@@ -60,11 +60,6 @@ import { upgradeExecutorFetchPrivilegedAccounts } from '../upgradeExecutorFetchP
 import { getBridgeUiConfigSchema, getBridgeUiConfigTransform } from './schemas/getBridgeUiConfig';
 import { getBridgeUiConfig } from '../getBridgeUiConfig';
 import {
-  getChainDeploymentInfoSchema,
-  getChainDeploymentInfoTransform,
-} from './schemas/getChainDeploymentInfo';
-import { getChainDeploymentInfo } from '../getChainDeploymentInfo';
-import {
   isTokenBridgeDeployedSchema,
   isTokenBridgeDeployedTransform,
 } from './schemas/isTokenBridgeDeployed';
@@ -154,11 +149,6 @@ import {
   schema as createTokenBridgeExampleSchema,
   execute as createTokenBridgeExampleExecute,
 } from './examples/createTokenBridge';
-import {
-  schema as getChainDeploymentInfoExampleSchema,
-  execute as getChainDeploymentInfoExampleExecute,
-} from './examples/getChainDeploymentInfo';
-
 describe('schema coverage', () => {
   it('getValidators', async () => {
     await assertSchemaCoverage(
@@ -290,14 +280,6 @@ describe('schema coverage', () => {
     await assertSchemaCoverage(
       getBridgeUiConfigSchema.transform(getBridgeUiConfigTransform),
       getBridgeUiConfig,
-      mocks,
-    );
-  });
-
-  it('getChainDeploymentInfo', async () => {
-    await assertSchemaCoverage(
-      getChainDeploymentInfoSchema.transform(getChainDeploymentInfoTransform),
-      getChainDeploymentInfo,
       mocks,
     );
   });
@@ -546,13 +528,5 @@ describe('schema coverage', () => {
         nativeToken: '0x0000000000000000000000000000000000000000',
       }),
     });
-  });
-
-  it('getChainDeploymentInfo example', async () => {
-    await assertSchemaCoverage(
-      getChainDeploymentInfoExampleSchema,
-      getChainDeploymentInfoExampleExecute,
-      mocks,
-    );
   });
 });
